Title: 최솟값 3개, Time: 176ms, Memory: 1MB, Status: Passed - CodeTree

+++ 241014/최솟값 3개/top-3-smallest-number.cpp
@@ -22,7 +22,7 @@ int main() {
             int n2 = -pq.top();
             pq.pop();
 
-            int ans = -pq.top();
+            long long ans = -pq.top();
 
             ans = ans*n1*n2;
             pq.push(-n1);
